Time: 0 ms (100%), Space: 7.7 MB (61.16%) - LeetHub

+++ 0063-unique-paths-ii/0063-unique-paths-ii.cpp
@@ -65,7 +65,7 @@ public:
     int uniquePathsWithObstacles(vector<vector<int>>& obstacleGrid) {
         int m = obstacleGrid.size();
         int n = obstacleGrid[0].size();
-        vector<vector<long long>> dp(m,vector<long long>(n,0));
+        vector<vector<int>> dp(m,vector<int>(n,0));
         if(obstacleGrid[m-1][n-1] == 1 || obstacleGrid[0][0] == 1)
             return 0;
         
@@ -77,8 +77,8 @@ public:
                     continue;
                 }
             
-                long long l = 0;
-                long long up = 0;
+                int l = 0;
+                int up = 0;
                 if(i>0 && obstacleGrid[i][j] == 0)
                     l = dp[i-1][j];
                 if(j>0 && obstacleGrid[i][j] == 0)
